debian/extra/network: use NamePolicy=mac only when ID_NET_NAME_MAC is set
After the kernel patch,
https://lore.kernel.org/linux-usb/20240605153340.25694-1-gmazyland@gmail.com/
even a real interface may not have persistent MAC address, and thus
ID_NET_NAME_MAC property may not be set on the device.

The 'mac' naming policy requires that ID_NET_NAME_MAC property is set on
the device, and if it is not, even if the .link file is applied to the
device, the interface is not renamed and it stays with the default name,
e.g. ethX. So, applying the .link file is mostly meaningless in such case.

Let's apply the .link file only when the interface has a persistent MAC
address, and thus ID_NET_NAME_MAC property is set.

C.f. https://github.com/systemd/systemd/issues/33104.