Drop redundant pot build
This was added many years ago, when apparently the upstream pot
generation wasn't run or wasn't working. This is not the case
anymore, pot files are regenerated upstream and checked in on
every release, so this manual step just updates the timestamp
in the existing template and nothing else. Drop it.