initramfs-tools: ensure rules file exists before invoking chzdev
If no files exist matching the glob, then the loop just iterates over
the string literal "/etc/udev/rules.d/*.rules". If this is passed to
chzdev, it fails with "no such file or directory."

Keep the glob, but skip the loop body if $rules does not exist.

Gbp-Dch: Ignore