systemd-boot: provide integration with shim
When installing/updating systemd-boot, if _both_ signed shim and signed
systemd-boot are available, and signed grub is not available, then take
over the ESP and the EFI boot entry, install systemd-boot as the second
stage loader, and install signed shim as the first stage loader and add
a boot entry named after the OS vendor pointing to it, and mark it as
the default.

When triggered, given dpkg does not tell us whether the trigger added
or removed a file, if signed grub is available then do nothing and
assume it will take over the ESP. If it is not, and either signed shim
or signed systemd-boot or both are not available, then remove all of
them from the ESP and the EFI boot entry. If both signed shim and
signed systemd-boot are available, then update them in the ESP.

Also documented at:

https://wiki.debian.org/SecureBoot#Secure_Boot_setup_with_systemd-boot