d/t/control: add more packages to dummy hint-testsuite-triggers
A regression was introduced by nvme-cli in testing because the testsuite
doesn't depend on it anymore directly, as mkosi builds it in the image.
Add back all the tools/libraries that are only part of the mkosi guest
image, so that autopkgtest triggers when they get updated, and
regressions can be flagged before they migrate to testing.