d/t/upstream: do not use --verbose with 'meson test'
Using verbose means the logs from all tests is printed even if they pass, which
causes the full test run log to be truncated on debci as it is too large.
Remove --verbose, as we anyway want to see the output only when a test case
fails, which happens without --verbose too.