udev: link to libsystemd-shared when building with noudeb profile
This was disabled because there was no libsystemd-shared package, so
it would have meant a dependency on the systemd package. But now we
have the package, so we can link against it.

Except there are udebs too, and that would mean creating a libsystemd-shared
udeb. So only change this for the noudeb profile for now.

Shrinks binary from 1.6M down to 611K.

Gbp-Dch: ignore