systemd-boot: use boot entry argument instead of installing as grub.efi on ESP
It turns out Shim can load sd-boot if the boot entry is set correctly.
As found by a user on https://github.com/rhboot/shim/issues/688#issuecomment-2558230720
the entry needs a terminating NUL character in order to work.

Use that and avoid installing sd-boot as GRUB, and remove all handling
of grub in the ESP, leaving it to its own package.

With the next version of Shim we can also write a CSV file so that
we can be sure it works on all firmwares.