Drop systemd-resolved package
The ctte has declared that the way the systemd-resolved tool works
is incompatible with their decision to prioritize avahi in Debian.
Furthermore, the resolved tool is being used to inflict pain on the
maintainer, and induce burnout.

Regrettably, the only safe solution to ensure this package is
compliant with this decision is to drop it, as
all reasonable alternatives put forward have been rejected:
https://salsa.debian.org/systemd-team/systemd/-/merge_requests/289

Closes: #1098914