Drop systemd-nspawn on arm64
The ctte has declared that the way the systemd-nspawn binary works
on arm64 is incompatible with their decision to leave in place a
recently introduced change in the debian-specific base-files
package, which caused a dpkg regression.
Furthermore, the nspawn tool is being used to inflict pain on the
maintainer, and induce burnout.

Regrettably, the only safe solution to ensure this package is
compliant with this decision is to drop the binary on arm64, as
all reasonable alternatives put forward have been rejected:
https://salsa.debian.org/systemd-team/systemd/-/merge_requests/289

Closes: #1079329